refactor controllers vs models
