+++ music_assistant/controllers/media/__init__.py
@@ -1,0 +1,1 @@
+"""Package with Media controllers."""
