+++ music_assistant/controllers/media/tracks.py
@@ -4,12 +4,12 @@ from __future__ import annotations
 import asyncio
 from typing import List, Optional, Union
 
+from music_assistant.controllers.database import TABLE_TRACKS
 from music_assistant.helpers.compare import (
     compare_artists,
     compare_track,
     loose_compare_strings,
 )
-from music_assistant.helpers.database import TABLE_TRACKS
 from music_assistant.helpers.json import json_serializer
 from music_assistant.models.enums import (
     EventType,
@@ -22,7 +22,6 @@ from music_assistant.models.errors import (
     UnsupportedFeaturedException,
 )
 from music_assistant.models.event import MassEvent
-from music_assistant.models.media_controller import MediaControllerBase
 from music_assistant.models.media_items import (
     Album,
     Artist,
@@ -30,6 +29,8 @@ from music_assistant.models.media_items import (
     Track,
     TrackAlbumMapping,
 )
+
+from .base import MediaControllerBase
 
 
 class TracksController(MediaControllerBase[Track]):
