+++ music_assistant/controllers/metadata/fanarttv.py
@@ -7,10 +7,10 @@ from typing import TYPE_CHECKING, Optional
 import aiohttp
 from asyncio_throttle import Throttler
 
+from music_assistant.controllers.cache import use_cache
 from music_assistant.helpers.app_vars import (  # pylint: disable=no-name-in-module
     app_var,
 )
-from music_assistant.helpers.cache import use_cache
 from music_assistant.models.media_items import (
     Album,
     Artist,
