+++ music_assistant/music_providers/__init__.py
@@ -1,1 +1,1 @@
-"""Package with Music Providers."""
+"""Package with Music Provider controllers."""
