+++ music_assistant/music_providers/filesystem/__init__.py
@@ -1,0 +1,3 @@
+"""Package with FileSystem Music provider(s)."""
+
+from .filesystem import FileSystemProvider  # noqa
