+++ music_assistant/music_providers/qobuz/__init__.py
@@ -1,0 +1,3 @@
+"""Package with Qobuz Music provider."""
+
+from .qobuz import QobuzProvider  # noqa
