+++ music_assistant/music_providers/spotify/__init__.py
@@ -1,0 +1,3 @@
+"""Package with Spotify Music provider."""
+
+from .spotify import SpotifyProvider  # noqa
