+++ music_assistant/music_providers/tunein/__init__.py
@@ -1,0 +1,3 @@
+"""Package with Tune-In Music provider."""
+
+from .tunein import TuneInProvider  # noqa
