+++ music_assistant/music_providers/url/__init__.py
@@ -1,0 +1,3 @@
+"""Package with URL Music provider."""
+
+from .url import URLProvider  # noqa
